Almost everything is Done, Business Logic absolutely fixed, hust edge cases needed  to test

+++ Domain/Services/IPaymentService.cs
@@ -4,8 +4,8 @@ namespace Domain.Services;
 
 public interface IPaymentService
 {
-    Task MakePayment(int amount, Guid loadId);
-    Task MakePenaltyPayment(int amount, Guid penaltyId);
+    Task MakePayment(decimal amount, Guid loadId);
+    Task MakePenaltyPayment(decimal amount, Guid penaltyId);
 
     Task<IEnumerable<Payment>> GetAllPayments(Client client);
 }
